move format function to utils
